alsa: rework eventfd handling
Make one function that updates the eventfd based on the state of
the plugin.

Do this check before getting the desciptors. The functional difference
is that the eventfd could become blocking now as well when getting the
descriptors. This fixes a problem where the poll would wake up without
any work to do.

See #1697