alsa-plugin: unblock alsa poll when there is pw error
When there is a core error, update_active method is called from
on_core_error. update_active is supposed to unblock alsa thread,
in case the alsa is waiting for more some room for data.

Checking for active state is not sufficient and spa_system_eventfd_write
needs to be called also in case of error.

Signed-off-by: Martin Geier <martin.geier@streamunlimited.com>