impl-node: Don't schedule non-active node
Avoid scheduling non-active nodes by removing the eventfd source from
the data loop when the node is deactivated. Read any old value from the
eventfd when starting the node again in case it has been written to
while the source was removed from the loop.

The ready callback can still be called when the node isn't active since
it is not called via the eventfd.