impl-node: activate links immediately
Don't wait for the node to be added to the graph before we activate
the links to it.

We don't do the reverse for shutdown and the activation counters won't
actually be updated until the node is added.

Waiting can cause race conditions in some specific cases (see in
screen sharing unit test) because the driver node can start
pushing buffers before the link has sent the Buffer io area to the
ports.

With this fix, the receiver (input stream) will first trigger
the input link activation, then the Start command and then it will be
added to the graph.

This does not entirely fix the race conditions when starting. Ideally,
the driver node should wait until all pending Start commands of the
nodes in the graph have completed.